moved autogen integration to its own dedicated package
