+++ build-dotnet.cmd
@@ -2,5 +2,5 @@
 set OUTDIR=artifacts\csharp
 rmdir /S /Q "%OUTDIR%"
 cargo build --release --manifest-path strathweb-phi-engine/Cargo.toml
-dotnet build packages\csharp\Strathweb.Phi.Engine -c Release
-dotnet pack packages\csharp\Strathweb.Phi.Engine -c Release -o %OUTDIR%
+dotnet build packages\csharp -c Release
+dotnet pack packages\csharp -c Release -o %OUTDIR%
