+++ packages/csharp/Strathweb.Phi.Engine.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Strathweb.Phi.Engine", "Strathweb.Phi.Engine\Strathweb.Phi.Engine.csproj", "{C10B0AA2-5D6B-4797-BB72-E0B361A22540}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Strathweb.Phi.Engine.AutoGen", "Strathweb.Phi.Engine.AutoGen\Strathweb.Phi.Engine.AutoGen.csproj", "{F111443F-651B-4810-A1D5-AA7CB30E6ABC}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C10B0AA2-5D6B-4797-BB72-E0B361A22540}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C10B0AA2-5D6B-4797-BB72-E0B361A22540}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C10B0AA2-5D6B-4797-BB72-E0B361A22540}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C10B0AA2-5D6B-4797-BB72-E0B361A22540}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F111443F-651B-4810-A1D5-AA7CB30E6ABC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F111443F-651B-4810-A1D5-AA7CB30E6ABC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F111443F-651B-4810-A1D5-AA7CB30E6ABC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F111443F-651B-4810-A1D5-AA7CB30E6ABC}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
