+++ samples/csharp/autogen/autogen.console.csproj
@@ -7,10 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-      <PackageReference Include="Strathweb.Phi.Engine" Version="0.1.1" />
-      <PackageReference Include="AutoGen" Version="0.1.0" />
+      <PackageReference Include="Strathweb.Phi.Engine.AutoGen" Version="0.1.1" />
       <PackageReference Include="FluentAssertions" Version="6.12.0" />
-      <PackageReference Include="System.Text.Json" Version="8.0.4" />
       <PackageReference Include="spectre.console" Version="0.47.0" />
   </ItemGroup>
 </Project>
